+++ srcjs/src/components/controls/ColumnDragHandle.svelte
@@ -65,7 +65,8 @@
   }
 </script>
 
-<!-- svelte-ignore a11y_no_static_element_interactions a11y_click_events_have_key_events -->
+<!-- svelte-ignore a11y_no_static_element_interactions -->
+<!-- svelte-ignore a11y_click_events_have_key_events -->
 <span
   class="drag-handle column-drag-handle"
   title="Drag to reorder"